# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Informasi Manajemen Inventaris Sekolah API",
        "description": "REST API untuk Sistem Informasi Manajemen Inventaris Sekolah. Mendukung modul Auth, Pengguna, Peran & Akses, Master Sekolah, Master Barang, Manajemen Aset, Transaksi Peminjaman & Permintaan, Pemeliharaan & Akhir Siklus Aset, serta Manajemen Database.",
        "contact": {
            "name": "Admin Sekolah",
            "email": "admin@sekolah.sch.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/pengguna": {
            "get": {
                "tags": [
                    "Pengguna"
                ],
                "summary": "Daftar semua pengguna",
                "description": "Tampilkan daftar semua pengguna.",
                "operationId": "indexPengguna",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PenggunaListResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pengguna"
                ],
                "summary": "Tambah pengguna baru",
                "description": "Simpan pengguna baru.",
                "operationId": "storePengguna",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePenggunaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PenggunaSingleResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/pengguna/{id}": {
            "get": {
                "tags": [
                    "Pengguna"
                ],
                "summary": "Detail pengguna",
                "description": "Tampilkan detail satu pengguna.",
                "operationId": "showPengguna",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PenggunaSingleResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pengguna"
                ],
                "summary": "Update pengguna",
                "description": "Update data pengguna.",
                "operationId": "updatePengguna",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePenggunaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PenggunaSingleResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pengguna"
                ],
                "summary": "Hapus pengguna",
                "description": "Hapus pengguna.",
                "operationId": "destroyPengguna",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PenggunaDeleteResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PenggunaResource": {
                "properties": {
                    "id_pengguna": {
                        "type": "integer",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "admin"
                    },
                    "id_peran": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "peran": {
                        "type": "object",
                        "nullable": true
                    },
                    "id_kelas": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "kelas": {
                        "type": "object",
                        "nullable": true
                    },
                    "id_mapel": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "mapel": {
                        "type": "object",
                        "nullable": true
                    },
                    "id_unit": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "unit": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "StorePenggunaRequest": {
                "required": [
                    "username",
                    "password",
                    "id_peran"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "guru01",
                        "maxLength": 100
                    },
                    "password": {
                        "type": "string",
                        "example": "password123",
                        "minLength": 8
                    },
                    "id_peran": {
                        "type": "integer",
                        "example": 2
                    },
                    "id_kelas": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "id_mapel": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "id_unit": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UpdatePenggunaRequest": {
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "guru01_updated",
                        "maxLength": 100
                    },
                    "password": {
                        "type": "string",
                        "example": "newpassword123",
                        "nullable": true,
                        "minLength": 8
                    },
                    "id_peran": {
                        "type": "integer",
                        "example": 2
                    },
                    "id_kelas": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "id_mapel": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "id_unit": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PenggunaListResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Daftar pengguna berhasil diambil."
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PenggunaResource"
                        }
                    }
                },
                "type": "object"
            },
            "PenggunaSingleResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Detail pengguna berhasil diambil."
                    },
                    "data": {
                        "$ref": "#/components/schemas/PenggunaResource"
                    }
                },
                "type": "object"
            },
            "PenggunaDeleteResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Pengguna berhasil dihapus."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}